add type definitions and standard return to several user/status/external api commands
