+++ e2e/cypress/support/api/index.js
@@ -3,6 +3,7 @@
 
 import './setup';
 import './preference';
+import './status';
 import './system';
 import './team';
 import './user';
