+++ e2e/cypress/support/api/status.d.ts
@@ -1,0 +1,34 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+/// <reference types="cypress" />
+
+// ***************************************************************
+// Each command should be properly documented using JSDoc.
+// See https://jsdoc.app/index.html for reference.
+// Basic requirements for documentation are the following:
+// - Meaningful description
+// - Specific link to https://api.mattermost.com
+// - Each parameter with `@params`
+// - Return value with `@returns`
+// - Example usage with `@example`
+// Custom command should follow naming convention of having `api` prefix, e.g. `apiLogin`.
+// ***************************************************************
+
+declare namespace Cypress {
+    interface Chainable<Subject = any> {
+
+        /**
+         * Update status of a current user.
+         * See https://api.mattermost.com/#tag/status/paths/~1users~1{user_id}~1status/put
+         * @param {String} status - "online" (default), "offline", "away" or "dnd"
+         * @returns {UserStatus} `out.status` as `UserStatus`
+         *
+         * @example
+         *   cy.apiUpdateUserStatus('offline').then(({status}) => {
+         *       // do something with status
+         *   });
+         */
+        apiUpdateUserStatus(status: string): Chainable<UserStatus>;
+    }
+}
